refactor: MVC 패턴에 따라 폴더 생성하여 파일 분류
------ file tree ------
src
 ┣ Controller
 ┃ ┣ BridgeGame.js
 ┃ ┣ ErrorHandling.js
 ┃ ┗ ViewController.js
 ┣ Model
 ┃ ┣ Bridge.js
 ┃ ┗ Map.js
 ┣ View
 ┃ ┣ InputView.js
 ┃ ┗ OutputView.js
 ┣ constants
 ┃ ┣ Command.js
 ┃ ┣ ErrorMessage.js
 ┃ ┗ Message.js
 ┣ App.js
 ┣ BridgeMaker.js
 ┗ BridgeRandomNumberGenerator.js
