+++ src/BridgeMaker.js
@@ -1,5 +1,5 @@
 //@ts-check
-const { UP, DOWN } = require('./Command');
+const { UP, DOWN } = require('./constants/Command');
 
 /**
  * 다리의 길이를 입력 받아서 다리를 생성해주는 역할을 한다.
